pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "evernote-backup"
version = "1.0.0"
description = "Backup & export Evernote notes and notebooks."
authors = ["vzhd1701 <vzhd1701@gmail.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/vzhd1701/evernote-backup"
repository = "https://github.com/vzhd1701/evernote-backup"
keywords = ["evernote", "yinxiang", "backup", "export", "enex"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Natural Language :: English",
    "Environment :: Console",
    "Operating System :: MacOS",
    "Operating System :: Microsoft",
    "Operating System :: Unix",
    "Topic :: Internet",
    "Topic :: System :: Archiving :: Backup",
    "Topic :: Terminals",
    "Topic :: Utilities",
]

[tool.poetry.scripts]
evernote-backup = "evernote_backup.cli:main"

[tool.poetry.dependencies]
python = "^3.6.1"
evernote3 = "^1.25.14"
xmltodict = "^0.12.0"
click = "^7.1.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.3"
pytest-cov = "^2.11.1"
pytest-mock = "^3.5.1"
black = "^20.8b1"
isort = "^5.8.0"
pre-commit = "^2.12.0"
commitizen = "^2.17.2"
flakehell = "^0.9.0"
wemake-python-styleguide = "^0.15.2"
xmltodict = "^0.12.0"

[tool.commitizen]
name = "cz_conventional_commits"
version = "1.0.0"
tag_format = "$version"
version_files = [
  "pyproject.toml:version",
  "evernote_backup/version.py"
]

[tool.flakehell]
base = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/master/styles/flakehell.toml"
max_line_length = 88
exclude = [
    "tests/*.py",
    "evernote_backup/evernote_client_classes.py",
]

[tool.flakehell.plugins]
# Maybe disable
# WPS211 too many arguments
# WPS214 too many methods
# WPS602 forbids @staticmethod
# E800 Found commented out code
# WPS306 forces to use class ClassName(object) <- purely for consistency

# Disabled [personal preference]
# WPS111 short names
# WPS305 forbids f-strings
# WPS326 implicit line concat
# Docstings, don't have any
# WPS226 String constant overuse

# Exceptions [false positives & formatting]
# WPS323 Found `%` string formatting
# C812 trailing comma <- handled by black
# E203 whitespace before : <- handled by black
# W503 Line break occurred before a binary operator

# E800 TEMPORARILY DISABLE COMMENTED CODE
#flake8-eradicate = ["+*", "-E800"]

pycodestyle = ["+*", "-E203", "-W503"]
flake8-commas = ["+*", "-C812"]
flake8-quotes = ["+*", "-Q000"]
flake8-rst-docstrings = ["-*"]
flake8-docstrings = ["-*"]
flake8-darglint = ["-*"]
wemake-python-styleguide = ["+*", "-WPS111", "-WPS323", "-WPS326", "-WPS305", "-WPS226"]

[tool.flakehell.exceptions."evernote_backup/cli_app_util.py"]
# Lots of tiny helper modules
wemake-python-styleguide = ["-WPS202"]

[tool.flakehell.exceptions."evernote_backup/evernote_client_util.py"]
# Complex decorator
wemake-python-styleguide = ["-WPS231", "-WPS232"]

[tool.flakehell.exceptions."evernote_backup/cli.py"]
# Too many arguments, argv options
wemake-python-styleguide = ["-WPS211"]

[tool.flakehell.exceptions."evernote_backup/cli_app.py"]
# Too many arguments, argv options
wemake-python-styleguide = ["-WPS211"]

[tool.flakehell.exceptions."evernote_backup/version.py"]
# I like it this way
wemake-python-styleguide = ["-WPS410"]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
skip_gitignore = true

[tool.coverage]
    [tool.coverage.report]
    show_missing = true
    exclude_lines = [
        # Have to re-enable the standard pragma
        'pragma: no cover',

        # Don't complain about missing debug-only code:
        'def __repr__',
        'if self\.debug',

        # Don't complain if tests don't hit defensive assertion code:
        'raise AssertionError',
        'raise NotImplementedError',

        # Don't complain if non-runnable code isn't run:
        'if 0:',
        'if __name__ == .__main__.:'
    ]
    omit = [
        '*/tests/*'
    ]
